[refactor] ProjectPost 패키지 컨벤션 맞추기
