+++ module-api/src/main/java/com/welcommu/moduleapi/projectpost/ProjectPostController.java
@@ -3,7 +3,7 @@ package com.welcommu.moduleapi.projectpost;
 import com.welcommu.modulecommon.dto.ApiResponse;
 import com.welcommu.moduleservice.projectpost.dto.ProjectPostListResponse;
 import com.welcommu.moduleservice.projectpost.dto.ProjectPostRequest;
-import com.welcommu.moduleservice.projectpost.service.ProjectPostService;
+import com.welcommu.moduleservice.projectpost.ProjectPostService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
