+++ module-service/src/main/java/com/welcommu/moduleservice/projectpost/ProjectPostService.java
@@ -1,4 +1,4 @@
-package com.welcommu.moduleservice.projectpost.service;
+package com.welcommu.moduleservice.projectpost;
 
 import com.welcommu.moduledomain.projectpost.ProjectPost;
 import com.welcommu.modulerepository.projectpost.repository.ProjectPostRepository;
